feat(behavior_path_planner): be able to switch framework by macro
