+++ planning/behavior_path_planner/CMakeLists.txt
@@ -15,6 +15,8 @@ set(common_src
   src/utilities.cpp
   src/path_utilities.cpp
   src/steering_factor_interface.cpp
+  src/behavior_path_planner_node.cpp
+  src/scene_module/scene_module_visitor.cpp
   src/turn_signal_decider.cpp
   src/util/avoidance/util.cpp
   src/util/lane_change/util.cpp
@@ -40,9 +42,7 @@ set(common_src
 
 if(USE_BT)
   ament_auto_add_library(behavior_path_planner_node SHARED
-    src/behavior_path_planner_node.cpp
     src/behavior_tree_manager.cpp
-    src/scene_module/scene_module_visitor.cpp
     src/scene_module/scene_module_bt_node_interface.cpp
     src/scene_module/side_shift/side_shift_module.cpp
     src/scene_module/avoidance/avoidance_module.cpp
@@ -60,12 +60,7 @@ if(USE_BT)
 
 else()
   ament_auto_add_library(behavior_path_planner_node SHARED
-    src/behavior_path_planner_node.cpp
     src/planner_manager.cpp
-    src/scene_module/scene_module_visitor.cpp
-    # src/scene_module_v2/avoidance
-    # src/scene_module_v2/lane_change
-    # src/scene_module_v2/...
     ${common_src}
   )
